Reorganized modules in golem-common
