+++ golem-cli/src/service/worker.rs
@@ -27,7 +27,7 @@ use golem_client::model::{
     InvokeParameters, InvokeResult, ScanCursor, StringFilterComparator, Type, WorkerFilter,
     WorkerNameFilter,
 };
-use golem_common::precise_json::PreciseJson;
+use golem_common::model::precise_json::PreciseJson;
 use golem_wasm_rpc::protobuf::type_annotated_value::TypeAnnotatedValue;
 use golem_wasm_rpc::type_annotated_value_from_str;
 use serde_json::Value;
