+++ golem-common/src/lib.rs
@@ -14,15 +14,12 @@
 
 pub mod cache;
 pub mod client;
-pub mod component_metadata;
 pub mod config;
 
-pub mod exports;
 pub mod grpc;
 pub mod metrics;
 pub mod model;
 pub mod newtype;
-pub mod precise_json;
 pub mod redis;
 pub mod retriable_error;
 pub mod retries;
