+++ golem-component-service-base/src/model.rs
@@ -1,4 +1,4 @@
-use golem_common::component_metadata::ComponentMetadata;
+use golem_common::model::component_metadata::ComponentMetadata;
 use golem_service_base::model::{
     ComponentName, ProtectedComponentId, UserComponentId, VersionedComponentId,
 };
