+++ golem-worker-executor-base/src/worker.rs
@@ -36,7 +36,7 @@ use crate::services::{
 use crate::workerctx::WorkerCtx;
 use anyhow::anyhow;
 use golem_common::config::RetryConfig;
-use golem_common::exports;
+use golem_common::model::exports;
 use golem_common::model::oplog::{
     OplogEntry, OplogIndex, TimestampedUpdateDescription, UpdateDescription, WorkerError,
     WorkerResourceId,
@@ -1520,94 +1520,94 @@ impl RunningWorker {
                                         target_version = target_version.to_string()
                                     );
                                     let do_break = async {
-                                    let _idempotency_key = {
-                                        let ctx = store.data_mut();
-                                        let idempotency_key = IdempotencyKey::fresh();
-                                        ctx.set_current_idempotency_key(idempotency_key.clone())
+                                        let _idempotency_key = {
+                                            let ctx = store.data_mut();
+                                            let idempotency_key = IdempotencyKey::fresh();
+                                            ctx.set_current_idempotency_key(idempotency_key.clone())
+                                                .await;
+                                            idempotency_key
+                                        };
+                                        store.data_mut().begin_call_snapshotting_function();
+                                        let result = invoke_worker(
+                                            "golem:api/save-snapshot@0.2.0.{save}".to_string(),
+                                            vec![],
+                                            store,
+                                            &instance,
+                                            CallingConvention::Component,
+                                            true,
+                                        )
                                             .await;
-                                        idempotency_key
-                                    };
-                                    store.data_mut().begin_call_snapshotting_function();
-                                    let result = invoke_worker(
-                                        "golem:api/save-snapshot@0.2.0.{save}".to_string(),
-                                        vec![],
-                                        store,
-                                        &instance,
-                                        CallingConvention::Component,
-                                        true,
-                                    )
-                                        .await;
-                                    store.data_mut().end_call_snapshotting_function();
+                                        store.data_mut().end_call_snapshotting_function();
 
-                                    match result {
-                                        Ok(InvokeResult::Succeeded { output, .. }) =>
-                                            if let Some(bytes) = Self::decode_snapshot_result(output) {
-                                                match store
-                                                    .data_mut()
-                                                    .get_public_state()
-                                                    .oplog()
-                                                    .create_snapshot_based_update_description(
-                                                        target_version,
-                                                        &bytes,
-                                                    )
-                                                    .await
-                                                {
-                                                    Ok(update_description) => {
-                                                        // Enqueue the update
-                                                        parent.enqueue_update(update_description).await;
+                                        match result {
+                                            Ok(InvokeResult::Succeeded { output, .. }) =>
+                                                if let Some(bytes) = Self::decode_snapshot_result(output) {
+                                                    match store
+                                                        .data_mut()
+                                                        .get_public_state()
+                                                        .oplog()
+                                                        .create_snapshot_based_update_description(
+                                                            target_version,
+                                                            &bytes,
+                                                        )
+                                                        .await
+                                                    {
+                                                        Ok(update_description) => {
+                                                            // Enqueue the update
+                                                            parent.enqueue_update(update_description).await;
 
-                                                        // Make sure to update the pending updates queue
-                                                        store.data_mut().update_pending_updates().await;
+                                                            // Make sure to update the pending updates queue
+                                                            store.data_mut().update_pending_updates().await;
 
-                                                        // Reactivate the worker
-                                                        final_decision = RetryDecision::Immediate;
+                                                            // Reactivate the worker
+                                                            final_decision = RetryDecision::Immediate;
 
-                                                        // Stop processing the queue to avoid race conditions
-                                                        true
+                                                            // Stop processing the queue to avoid race conditions
+                                                            true
+                                                        }
+                                                        Err(error) => {
+                                                            Self::fail_update(target_version, format!("failed to store the snapshot for manual update: {error}"), store).await;
+                                                            false
+                                                        }
                                                     }
-                                                    Err(error) => {
-                                                        Self::fail_update(target_version, format!("failed to store the snapshot for manual update: {error}"), store).await;
-                                                        false
-                                                    }
-                                                }
-                                            } else {
-                                                Self::fail_update(target_version, "failed to get a snapshot for manual update: invalid snapshot result".to_string(), store).await;
+                                                } else {
+                                                    Self::fail_update(target_version, "failed to get a snapshot for manual update: invalid snapshot result".to_string(), store).await;
+                                                    false
+                                                },
+                                            Ok(InvokeResult::Failed { error, .. }) => {
+                                                Self::fail_update(
+                                                    target_version,
+                                                    format!("failed to get a snapshot for manual update: {error}"),
+                                                    store,
+                                                ).await;
                                                 false
-                                            },
-                                        Ok(InvokeResult::Failed { error, .. }) => {
-                                            Self::fail_update(
-                                                target_version,
-                                                format!("failed to get a snapshot for manual update: {error}"),
-                                                store,
-                                            ).await;
-                                            false
+                                            }
+                                            Ok(InvokeResult::Exited { .. }) => {
+                                                Self::fail_update(
+                                                    target_version,
+                                                    "failed to get a snapshot for manual update: it called exit".to_string(),
+                                                    store,
+                                                ).await;
+                                                false
+                                            }
+                                            Ok(InvokeResult::Interrupted { interrupt_kind, .. }) => {
+                                                Self::fail_update(
+                                                    target_version,
+                                                    format!("failed to get a snapshot for manual update: {interrupt_kind:?}"),
+                                                    store,
+                                                ).await;
+                                                false
+                                            }
+                                            Err(error) => {
+                                                Self::fail_update(
+                                                    target_version,
+                                                    format!("failed to get a snapshot for manual update: {error:?}"),
+                                                    store,
+                                                ).await;
+                                                false
+                                            }
                                         }
-                                        Ok(InvokeResult::Exited { .. }) => {
-                                            Self::fail_update(
-                                                target_version,
-                                                "failed to get a snapshot for manual update: it called exit".to_string(),
-                                                store,
-                                            ).await;
-                                            false
-                                        }
-                                        Ok(InvokeResult::Interrupted { interrupt_kind, .. }) => {
-                                            Self::fail_update(
-                                                target_version,
-                                                format!("failed to get a snapshot for manual update: {interrupt_kind:?}"),
-                                                store,
-                                            ).await;
-                                            false
-                                        }
-                                        Err(error) => {
-                                            Self::fail_update(
-                                                target_version,
-                                                format!("failed to get a snapshot for manual update: {error:?}"),
-                                                store,
-                                            ).await;
-                                            false
-                                        }
-                                    }
-                                }.instrument(span).await;
+                                    }.instrument(span).await;
                                     if do_break {
                                         break;
                                     }
@@ -1732,7 +1732,7 @@ impl InvocationResult {
                 OplogEntry::Interrupted { .. } => Err(TrapType::Interrupt(InterruptKind::Interrupt)),
                 OplogEntry::Exited { .. } => Err(TrapType::Exit),
                 _ => panic!("Unexpected oplog entry pointed by invocation result at index {oplog_idx} for {oplog:?}")
-            } ;
+            };
 
             *self = Self::Cached { result, oplog_idx }
         }
