+++ golem-worker-service-base/src/evaluator/internal.rs
@@ -4,8 +4,8 @@ use crate::evaluator::EvaluationContext;
 use crate::evaluator::EvaluationError;
 use crate::primitive::GetPrimitive;
 use crate::worker_bridge_execution::{RefinedWorkerResponse, WorkerRequest, WorkerRequestExecutor};
+use golem_common::model::precise_json::PreciseJson;
 use golem_common::model::{ComponentId, IdempotencyKey};
-use golem_common::precise_json::PreciseJson;
 use golem_wasm_ast::analysis::AnalysedType;
 use golem_wasm_rpc::protobuf::type_annotated_value::TypeAnnotatedValue;
 use golem_wasm_rpc::protobuf::typed_result::ResultValue;
